add cross platform winit/pixels renderer, add more documentation, and move into a single crate
