--- cli/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "wasmstation-cli"
-version = "0.1.0"
-authors.workspace = true
-license.workspace = true
-edition.workspace = true
-
-[dependencies]
-wasmstation-core = { path = "../core/" }
-wasmstation-desktop = { path = "../desktop/" }
-wasmstation-wasmer = { path = "../wasmer/" }
-wasmstation-wasmi = { path = "../wasmi/" }
-
-argh = "0.1"
-pretty_env_logger = "0.4"
-
-[[bin]]
-name = "wasmstation"
-path = "src/main.rs"
