--- cli/src/template/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "{crate_name}"
-version = "0.1.0"
-edition = "2021"
-
-[profile.release]
-codegen-units = 1
-lto = true
-opt-level = "z"
-strip = true
-
-[build-dependencies]
-wasmer = "3.1"
-
-[dependencies]
-wasmstation_desktop = { git = "https://github.com/wasmstation/wasmstation", package = "wasmstation-desktop" }
-wasmstation_wasmer = { git = "https://github.com/wasmstation/wasmstation", package = "wasmstation-wasmer" }
-wasmstation_core = { git = "https://github.com/wasmstation/wasmstation", package = "wasmstation-core" }
