+++ examples/web-basic/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "web-basic"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+wasmstation = { path = "../..", default-features = false, features = ["wasmi", "gpu-renderer"] }
+
+log = "0.4"
+console_error_panic_hook = "0.1.7"
+console_log = "1.0"
+wasm-bindgen = "0.2"
