+++ examples/web-basic/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <title>Wasmstation</title>
+</head>
+
+<body>
+  <h1>Wasmstation</h1>
+  <canvas id="wasmstation" oncontextmenu="return false;"></canvas>
+</body>
+
+</html>
