+++ template/main.rs
@@ -1,0 +1,14 @@
+use std::env;
+
+use wasmstation::{WasmerBackend, Console, sdl2_renderer::launch_desktop};
+
+const WASM_BYTES: &[u8] = include_bytes!(concat!(env!("OUT_DIR"), "/wasm.module"));
+
+fn main() {
+    launch_desktop(
+        WasmerBackend::precompiled(WASM_BYTES, &Console::new(|s| println!("{s}"))).unwrap(),
+        &env::current_dir().unwrap(),
+        {window_scale},
+    )
+    .unwrap();
+}
