--- wasmi/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "wasmstation-wasmi"
-version = "0.1.0"
-authors.workspace = true
-license.workspace = true
-edition.workspace = true
-
-
-[dependencies]
-log.workspace = true
-bytemuck.workspace = true
-wasmi = { version = "0.29", default-features = false }
-wasmstation-core = { path = "../core/" }
